:art: moved TextFilter to be its own top level entity
