+++ modules/components/src/TextFilter/index.js
@@ -1,0 +1,1 @@
+export { default, generateNextSQON } from './TextFilter';
